@@ -3,30 +3,28 @@
 namespace App\Controller\Admin;
 
 use App\Entity\Album;
-use App\Entity\Media;
 use App\Form\AlbumType;
-use App\Form\MediaType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
 
 class AlbumController extends AbstractController
 {
-    /**
-     * @Route("/admin/album", name="admin_album_index")
-     */
-    public function index()
-    {
+
+	#[Route(path: '/admin/album', name: 'admin_album_index')]
+	public function index(): Response
+	{
         $albums = $this->getDoctrine()->getRepository(Album::class)->findAll();
 
         return $this->render('admin/album/index.html.twig', ['albums' => $albums]);
     }
 
-    /**
-     * @Route("/admin/album/add", name="admin_album_add")
-     */
-    public function add(Request $request)
-    {
+
+	#[Route(path: '/admin/album/add', name: 'admin_album_add')]
+	public function add(Request $request): RedirectResponse|Response
+	{
         $album = new Album();
         $form = $this->createForm(AlbumType::class, $album);
         $form->handleRequest($request);
@@ -41,11 +39,10 @@ class AlbumController extends AbstractController
         return $this->render('admin/album/add.html.twig', ['form' => $form->createView()]);
     }
 
-    /**
-     * @Route("/admin/album/update/{id}", name="admin_album_update")
-     */
-    public function update(Request $request, int $id)
-    {
+
+	#[Route(path: '/admin/album/update/{id}', name: 'admin_album_update')]
+	public function update(Request $request, int $id): RedirectResponse|Response
+	{
         $album = $this->getDoctrine()->getRepository(Album::class)->find($id);
         $form = $this->createForm(AlbumType::class, $album);
         $form->handleRequest($request);
@@ -59,11 +56,10 @@ class AlbumController extends AbstractController
         return $this->render('admin/album/update.html.twig', ['form' => $form->createView()]);
     }
 
-    /**
-     * @Route("/admin/album/delete/{id}", name="admin_album_delete")
-     */
-    public function delete(int $id)
-    {
+
+	#[Route(path: '/admin/album/delete/{id}', name: 'admin_album_delete')]
+	public function delete(int $id): RedirectResponse
+	{
         $media = $this->getDoctrine()->getRepository(Album::class)->find($id);
         $this->getDoctrine()->getManager()->remove($media);
         $this->getDoctrine()->getManager()->flush();
