@@ -5,16 +5,17 @@ namespace App\Controller\Admin;
 use App\Entity\Media;
 use App\Form\MediaType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
+
 
 class MediaController extends AbstractController
 {
-    /**
-     * @Route("/admin/media", name="admin_media_index")
-     */
-    public function index(Request $request)
-    {
+	#[Route(path: '/admin/media', name: 'admin_media_index')]
+	public function index(Request $request): Response
+	{
         $page = $request->query->getInt('page', 1);
 
         $criteria = [];
@@ -38,11 +39,9 @@ class MediaController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/admin/media/add", name="admin_media_add")
-     */
-    public function add(Request $request)
-    {
+	#[Route(path: '/admin/media/add', name: 'admin_media_add')]
+	public function add(Request $request): RedirectResponse|Response
+	{
         $media = new Media();
         $form = $this->createForm(MediaType::class, $media, ['is_admin' => $this->isGranted('ROLE_ADMIN')]);
         $form->handleRequest($request);
@@ -62,11 +61,10 @@ class MediaController extends AbstractController
         return $this->render('admin/media/add.html.twig', ['form' => $form->createView()]);
     }
 
-    /**
-     * @Route("/admin/media/delete/{id}", name="admin_media_delete")
-     */
-    public function delete(int $id)
-    {
+
+	#[Route(path: '/admin/media/delete/{id}', name: 'admin_media_delete')]
+	public function delete(int $id): RedirectResponse
+	{
         $media = $this->getDoctrine()->getRepository(Media::class)->find($id);
         $this->getDoctrine()->getManager()->remove($media);
         $this->getDoctrine()->getManager()->flush();
