@@ -3,16 +3,15 @@
 namespace App\Controller\Admin;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 
 class SecurityController extends AbstractController
 {
-    /**
-     * @Route("/login", name="admin_login")
-     */
-    public function login(AuthenticationUtils $authenticationUtils)
-    {
+	#[Route(path: '/login', name: 'admin_login')]
+	public function login(AuthenticationUtils $authenticationUtils): Response
+	{
         $error = $authenticationUtils->getLastAuthenticationError();
         $lastUsername = $authenticationUtils->getLastUsername();
         return $this->render('admin/login.html.twig', [
